Remove locking; we don't need it
The only reason we put in locking was to keep from writing to a closed
channel. Now we have made the channel buffered and we don't bother
closing it. So, the logic is much simpler. And, we don't get stuck with
every go routine waiting on every other go routine to release the lock
just so it can check if it can early-exit.